// prisma/generated/zod/inputTypeSchemas/OrganizationUncheckedUpdateInputSchema.ts
export const OrganizationUncheckedUpdateInputSchema: z.ZodType<Prisma.OrganizationUncheckedUpdateInput> = z.object({
  id: z.union([ z.string().cuid(),z.lazy(() => StringFieldUpdateOperationsInputSchema) ]).optional(),
  name: z.union([ z.string(),z.lazy(() => StringFieldUpdateOperationsInputSchema) ]).optional(),
  subdomain: z.union([ z.string(),z.lazy(() => NullableStringFieldUpdateOperationsInputSchema) ]).optional().nullable(),
  logoUrl: z.union([ z.string(),z.lazy(() => NullableStringFieldUpdateOperationsInputSchema) ]).optional().nullable(),
  createdAt: z.union([ z.coerce.date(),z.lazy(() => DateTimeFieldUpdateOperationsInputSchema) ]).optional(),
  updatedAt: z.union([ z.coerce.date(),z.lazy(() => DateTimeFieldUpdateOperationsInputSchema) ]).optional(),
  memberships: z.lazy(() => MembershipUncheckedUpdateManyWithoutOrganizationNestedInputSchema).optional(),
  locations: z.lazy(() => LocationUncheckedUpdateManyWithoutOrganizationNestedInputSchema).optional(),
  roles: z.lazy(() => RoleUncheckedUpdateManyWithoutOrganizationNestedInputSchema).optional(),
  machines: z.lazy(() => MachineUncheckedUpdateManyWithoutOrganizationNestedInputSchema).optional(),
  issues: z.lazy(() => IssueUncheckedUpdateManyWithoutOrganizationNestedInputSchema).optional(),
  priorities: z.lazy(() => PriorityUncheckedUpdateManyWithoutOrganizationNestedInputSchema).optional(),
  issueStatuses: z.lazy(() => IssueStatusUncheckedUpdateManyWithoutOrganizationNestedInputSchema).optional(),
  collectionTypes: z.lazy(() => CollectionTypeUncheckedUpdateManyWithoutOrganizationNestedInputSchema).optional(),
  issueHistory: z.lazy(() => IssueHistoryUncheckedUpdateManyWithoutOrganizationNestedInputSchema).optional(),
  attachments: z.lazy(() => AttachmentUncheckedUpdateManyWithoutOrganizationNestedInputSchema).optional(),
  pinballMapConfig: z.lazy(() => PinballMapConfigUncheckedUpdateOneWithoutOrganizationNestedInputSchema).optional()
}).strict();